Creating login and register route and adding middleware to route

+++ routes/web.php
@@ -2,12 +2,14 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\JobController;
-
+use App\Http\Controllers\UserController;
 
 Route::get('/', [JobController::class, 'index']);
-Route::get('/jobs/create', [JobController::class, 'create']);
+Route::get('/jobs/create', [JobController::class, 'create'])->middleware('auth');
 Route::get('/jobs/{job}', [JobController::class, 'show']);
-Route::post('/jobs', [JobController::class, 'store']);
-Route::get('/jobs/{id}/edit', [JobController::class, 'edit']);
-Route::put('/jobs/{id}', [JobController::class, 'update']);
-Route::delete('/jobs/{job}', [JobController::class, 'delete']);
+Route::post('/jobs', [JobController::class, 'store'])->middleware('auth');
+Route::get('/jobs/{id}/edit', [JobController::class, 'edit'])->middleware('auth');
+Route::put('/jobs/{id}', [JobController::class, 'update'])->middleware('auth');
+Route::delete('/jobs/{job}', [JobController::class, 'delete'])->middleware('auth');
+Route::get('/login', [UserController::class, 'login'])->name('login');
+Route::get('/register', [UserController::class, 'register'])->name('register');
